+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.1.0
+
+* `Server` and related classes can now take an `onUnhandledError` callback to
+  notify callers of unhandled exceptions.
+
+## 2.0.10
+
+* Allow `stream_channel` version 2.x
+
 ## 2.0.8
 
 * Updated SDK version to 2.0.0-dev.17.0